+++ docs/js/datepicker.js
@@ -1,11 +1,11 @@
 /*!
- * Datepicker v1.0.4
+ * Datepicker v1.0.9
  * https://fengyuanchen.github.io/datepicker
  *
  * Copyright 2014-present Chen Fengyuan
  * Released under the MIT license
  *
- * Date: 2019-01-06T02:42:36.188Z
+ * Date: 2019-09-21T06:57:34.100Z
  */
 
 (function (global, factory) {
@@ -106,7 +106,9 @@
     pick: null
   };
 
-  var WINDOW = typeof window !== 'undefined' ? window : {};
+  var IS_BROWSER = typeof window !== 'undefined';
+  var WINDOW = IS_BROWSER ? window : {};
+  var IS_TOUCH_DEVICE = IS_BROWSER ? 'ontouchstart' in WINDOW.document.documentElement : false;
   var NAMESPACE = 'datepicker';
   var EVENT_CLICK = "click.".concat(NAMESPACE);
   var EVENT_FOCUS = "focus.".concat(NAMESPACE);
@@ -116,6 +118,7 @@
   var EVENT_RESIZE = "resize.".concat(NAMESPACE);
   var EVENT_SCROLL = "scroll.".concat(NAMESPACE);
   var EVENT_SHOW = "show.".concat(NAMESPACE);
+  var EVENT_TOUCH_START = "touchstart.".concat(NAMESPACE);
   var CLASS_HIDE = "".concat(NAMESPACE, "-hide");
   var LANGUAGES = {};
   var VIEWS = {
@@ -268,6 +271,11 @@
         $(window).on(EVENT_RESIZE, this.onResize = proxy(this.place, this));
         $(document).on(EVENT_CLICK, this.onGlobalClick = proxy(this.globalClick, this));
         $(document).on(EVENT_KEYUP, this.onGlobalKeyup = proxy(this.globalKeyup, this));
+
+        if (IS_TOUCH_DEVICE) {
+          $(document).on(EVENT_TOUCH_START, this.onTouchStart = proxy(this.touchstart, this));
+        }
+
         this.place();
       }
     },
@@ -289,6 +297,10 @@
         $(window).off(EVENT_RESIZE, this.onResize);
         $(document).off(EVENT_CLICK, this.onGlobalClick);
         $(document).off(EVENT_KEYUP, this.onGlobalKeyup);
+
+        if (IS_TOUCH_DEVICE) {
+          $(document).off(EVENT_TOUCH_START, this.onTouchStart);
+        }
       }
     },
     toggle: function toggle() {
@@ -498,20 +510,11 @@
         }
 
         if (parts.length === format.parts.length) {
+          // Set year and month first
           $.each(parts, function (i, part) {
             var value = parseInt(part, 10);
 
             switch (format.parts[i]) {
-              case 'dd':
-              case 'd':
-                date.setDate(value);
-                break;
-
-              case 'mm':
-              case 'm':
-                date.setMonth(value - 1);
-                break;
-
               case 'yy':
                 date.setFullYear(2000 + value);
                 break;
@@ -519,6 +522,24 @@
               case 'yyyy':
                 // Converts 2-digit year to 2000+
                 date.setFullYear(part.length === 2 ? 2000 + value : value);
+                break;
+
+              case 'mm':
+              case 'm':
+                date.setMonth(value - 1);
+                break;
+
+              default:
+            }
+          }); // Set day in the last to avoid converting `31/10/2019` to `01/10/2019`
+
+          $.each(parts, function (i, part) {
+            var value = parseInt(part, 10);
+
+            switch (format.parts[i]) {
+              case 'dd':
+              case 'd':
+                date.setDate(value);
                 break;
 
               default:
@@ -618,7 +639,7 @@
           if (format.hasMonth) {
             this.showView(VIEWS.MONTHS);
           } else {
-            $target.addClass(options.pickedClass).siblings().removeClass(options.pickedClass);
+            $target.siblings(".".concat(options.pickedClass)).removeClass(options.pickedClass).data('view', 'year');
             this.hideView();
           }
 
@@ -626,16 +647,17 @@
           break;
 
         case 'year':
-          viewYear = parseInt($target.text(), 10);
-          date.setFullYear(viewYear);
+          viewYear = parseInt($target.text(), 10); // Set date first to avoid month changing (#195)
+
           date.setDate(getMinDay(viewYear, viewMonth, viewDay));
-          viewDate.setFullYear(viewYear);
+          date.setFullYear(viewYear);
           viewDate.setDate(getMinDay(viewYear, viewMonth, viewDay));
+          viewDate.setFullYear(viewYear);
 
           if (format.hasMonth) {
             this.showView(VIEWS.MONTHS);
           } else {
-            $target.addClass(options.pickedClass).siblings().removeClass(options.pickedClass);
+            $target.addClass(options.pickedClass).data('view', 'year picked').siblings(".".concat(options.pickedClass)).removeClass(options.pickedClass).data('view', 'year');
             this.hideView();
           }
 
@@ -671,7 +693,7 @@
           if (format.hasDay) {
             this.showView(VIEWS.DAYS);
           } else {
-            $target.addClass(options.pickedClass).siblings().removeClass(options.pickedClass);
+            $target.siblings(".".concat(options.pickedClass)).removeClass(options.pickedClass).data('view', 'month');
             this.hideView();
           }
 
@@ -691,7 +713,7 @@
           if (format.hasDay) {
             this.showView(VIEWS.DAYS);
           } else {
-            $target.addClass(options.pickedClass).siblings().removeClass(options.pickedClass);
+            $target.addClass(options.pickedClass).data('view', 'month picked').siblings(".".concat(options.pickedClass)).removeClass(options.pickedClass).data('view', 'month');
             this.hideView();
           }
 
@@ -707,10 +729,13 @@
             viewMonth += 1;
           }
 
-          viewDay = parseInt($target.text(), 10);
+          viewDay = parseInt($target.text(), 10); // Set date to 1 to avoid month changing (#195)
+
+          date.setDate(1);
           date.setFullYear(viewYear);
           date.setMonth(viewMonth);
           date.setDate(viewDay);
+          viewDate.setDate(1);
           viewDate.setFullYear(viewYear);
           viewDate.setMonth(viewMonth);
           viewDate.setDate(viewDay);
@@ -761,6 +786,15 @@
 
       if (this.isInput && target !== this.element && this.shown && (key === 'Tab' || keyCode === 9)) {
         this.hide();
+      }
+    },
+    touchstart: function touchstart(_ref3) {
+      var target = _ref3.target;
+
+      // Emulate click in touch devices to support hiding the picker automatically (#197).
+      if (this.isInput && target !== this.element && !$.contains(this.$picker[0], target)) {
+        this.hide();
+        this.element.blur();
       }
     }
   };